refactor: 패키지 구조 수정 - 비즈니스 로직 수행은 domain 패키지 - 예외 메세지 설정은 config 패키지 - 유틸성 로직 수행은 utils 패키지에 위치
