+++ src/main/java/stringcalculator/domain/StringCalculator.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.domain;
 
 import java.util.Objects;
 import java.util.*;
