+++ src/test/java/stringcalculator/PositiveIntegerListTest.java
@@ -2,6 +2,7 @@ package stringcalculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import stringcalculator.domain.PositiveIntegerList;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
