+++ src/test/java/stringcalculator/StringCalculatorTest.java
@@ -3,6 +3,7 @@ package stringcalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import stringcalculator.domain.StringCalculator;
 
 import java.util.List;
